dsl modules renaming
