+++ chimney/src/main/scala/io/scalaland/chimney/OldDslMacros.scala
@@ -2,7 +2,7 @@ package io.scalaland.chimney
 
 import scala.language.experimental.macros
 
-private[chimney] object DslMacros {
+private[chimney] object OldDslMacros {
 
   def constFieldSelector(c: scala.reflect.macros.whitebox.Context)(selector: c.Tree, value: c.Tree): c.Tree = {
     import c.universe._
